Fix incorrect averageing of CPU usage
The CPU usage of a container is against the total cpu time on the
Host. Averaging it doesn't make sense.

The `docker stats` command multiplies it by number of CPUs to mimic
top command, which will result in a number greater than 100%,
for e.g. 400% for a 4 cpu system.

In our case we want to keep it within 100%, so multiplying is not
needed.